+++ CHANGES.rst
@@ -10,6 +10,10 @@ Added
 - Linters can now be installed and run in the GitHub Action using the ``lint:`` option.
 - Sort imports only if the range of modified lines overlaps with changes resulting from
   sorting the imports.
+- Allow force enabling/disabling of syntax highlighting using the ``color`` option in
+  ``pyproject.toml``, the ``PY_COLORS`` environment variable, and the
+  ``--color``/``--no-color`` command line options.
+- Syntax highlighting is now enabled by default in the GitHub Action.
 
 Fixed
 -----